createTextField module
